+++ y-collections/src/jvmTest/kotlin/com/github/whyrising/y/collections/KeywordTest.kt
@@ -1,7 +1,7 @@
 package com.github.whyrising.y.collections
 
-import com.github.whyrising.y.collections.concretions.map.m
 import com.github.whyrising.y.collections.concretions.set.hashSet
+import com.github.whyrising.y.collections.core.m
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.booleans.shouldBeFalse
 import io.kotest.matchers.booleans.shouldBeTrue