+++ y-collections/src/jvmTest/kotlin/com/github/whyrising/y/collections/SymbolTest.kt
@@ -1,7 +1,7 @@
 package com.github.whyrising.y.collections
 
-import com.github.whyrising.y.collections.concretions.map.m
 import com.github.whyrising.y.collections.concretions.set.hashSet
+import com.github.whyrising.y.collections.core.m
 import com.github.whyrising.y.collections.util.Murmur3
 import com.github.whyrising.y.collections.util.hashCombine
 import io.kotest.core.spec.style.FreeSpec